Added default formatstr (%g) to Axis to correct formatting bug. Fixed bug in NANMin and NANMax vars Fixed nan bugs in isnonzero() and difference(), corrected docstring error Forced double precision when converting values to dates in val_as_date

+++ pygeode/timeaxis.py
@@ -638,7 +638,7 @@ class CalendarTime(Time):
 
     iyear, imonth, iday, ihour, iminute, isecond = self._get_dates(startdate)
 
-    values = np.ascontiguousarray(np.round(vals * self.unitfactor[units]), dtype='int64')
+    values = np.ascontiguousarray(np.round(vals.astype('float64') * self.unitfactor[units]), dtype='int64')
     n = len(values)
     year   = np.empty(n, dtype='int32')
     month  = np.empty(n, dtype='int32')
